+++ packages/vscode-extension/src/chat-view-provider.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
-import axios from 'axios';
+import axios, { AxiosHeaders } from 'axios';
+import { RawAxiosRequestHeaders } from 'axios/index';
 
 export default class ChatViewProvider implements vscode.WebviewViewProvider {
 
@@ -7,14 +8,14 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
   private requestAbortController;
 
   constructor(
-    private readonly context: vscode.ExtensionContext
+    private readonly context: vscode.ExtensionContext,
   ) {
   }
 
   public resolveWebviewView(
     webviewView: vscode.WebviewView,
     context: vscode.WebviewViewResolveContext,
-    _token: vscode.CancellationToken
+    _token: vscode.CancellationToken,
   ) {
     this.webView = webviewView;
 
@@ -23,14 +24,18 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
       enableScripts: true,
 
       localResourceRoots: [
-        this.context.extensionUri
-      ]
+        this.context.extensionUri,
+      ],
     };
 
     webviewView.webview.html = this.getWebviewHtml(webviewView.webview);
 
     webviewView.webview.onDidReceiveMessage(data => {
       switch (data.type) {
+        case 'sendCommand': {
+          void this.sendPolyCommandRequest(data.value);
+          break;
+        }
         case 'sendQuestion': {
           void this.sendPolyQuestionRequest(data.value);
           break;
@@ -47,9 +52,6 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
     const apiBaseUrl = vscode.workspace.getConfiguration('poly').get('apiBaseUrl');
     const apiKey = vscode.workspace.getConfiguration('poly').get('apiKey');
 
-    console.log('Restarting TS server...');
-    vscode.commands.executeCommand('typescript.restartTsServer');
-
     if (!apiBaseUrl || !apiKey) {
       vscode.window.showErrorMessage('Please set the API base URL and API key in the extension settings.');
       return;
@@ -57,7 +59,7 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
 
     this.webView?.webview.postMessage({
       type: 'addQuestion',
-      value: message
+      value: message,
     });
     this.webView?.webview.postMessage({
       type: 'setLoading',
@@ -66,15 +68,18 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
     this.requestAbortController?.abort();
     this.requestAbortController = new AbortController();
     try {
-      const {data} = await axios.post(`${apiBaseUrl}/chat/question`, {
-        message
+      const { data } = await axios.post(`${apiBaseUrl}/chat/question`, {
+        message,
       }, {
+        headers: {
+          'X-PolyApiKey': apiKey,
+        } as RawAxiosRequestHeaders,
         signal: this.requestAbortController.signal,
       });
 
       this.webView?.webview.postMessage({
         type: 'addResponseTexts',
-        value: data.texts
+        value: data.texts,
       });
     } catch (error) {
       console.error(error);
@@ -82,16 +87,42 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
         type: 'addResponseTexts',
         value: [{
           type: 'error',
-          value: error.message
-        }]
+          value: error.message,
+        }],
       });
+    }
+  }
+
+  private async sendPolyCommandRequest(command: string) {
+    const apiBaseUrl = vscode.workspace.getConfiguration('poly').get('apiBaseUrl');
+    const apiKey = vscode.workspace.getConfiguration('poly').get('apiKey');
+
+    if (!apiBaseUrl || !apiKey) {
+      vscode.window.showErrorMessage('Please set the API base URL and API key in the extension settings.');
+      return;
+    }
+
+    this.webView?.webview.postMessage({
+      type: 'clearConversation',
+    });
+
+    try {
+      await axios.post(`${apiBaseUrl}/chat/command`, {
+        command,
+      }, {
+        headers: {
+          'X-PolyApiKey': apiKey,
+        } as RawAxiosRequestHeaders
+      });
+    } catch (error) {
+      console.error(error);
     }
   }
 
   focusMessageInput() {
     this.webView?.show();
     this.webView?.webview.postMessage({
-      type: 'focusMessageInput'
+      type: 'focusMessageInput',
     });
   }
 
@@ -120,20 +151,13 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
       </head>
       <body class='overflow-hidden'>
         <div class='flex flex-col h-screen'>
-          <div id='action-bar' class='flex justify-end hidden mb-2'>
-            <button id='clear-conversation-button' class='rounded-lg p-1.5 ml-2'>
-              <svg xmlns='http://www.w3.org/2000/svg' width='20' height='20' viewBox='0 0 128 128'>
-                <path fill='currentColor' d='M112.701 13.523c4.614 2.664 6.206 8.524 3.543 13.137L95.789 62.088l16.48 9.515a9.578 9.578 0 0 1 3.517 13.122l-5.59 9.683-8.031-4.636c-3.182 4.615-7.29 10.527-11.003 15.481-2.788 3.721-3.857 6.62-7.09 8.888-1.617 1.134-3.795 1.7-5.909 1.627-2.113-.072-4.282-.603-7.026-1.467a3.88 3.88 0 0 1-.439-.148L23.62 92.85c-2.944-1.331-5.304-2.305-7.094-3.046-1.79-.741-2.703-.945-4.133-1.932-.358-.246-.859-.466-1.462-1.649-.302-.591-.596-1.569-.402-2.576.193-1.008.814-1.802 1.287-2.23.947-.858 1.465-.904 1.88-1.015.416-.11.727-.151 1.058-.194 1.323-.171 2.915-.197 5.327-.319l12.162-.64c2.861-.145 6.935-3.557 9.122-6.77l8.683-12.798-5.117-2.954 5.59-9.684c2.664-4.613 8.535-6.164 13.148-3.5l15.457 8.924L99.58 17.039a9.578 9.578 0 0 1 13.121-3.516zm-17.254 72.37L56.79 63.575l-9.023 13.266c-2.979 4.377-7.96 9.794-15.136 10.156l-2.778.145 43.602 19.75c.031.01.036-.006.067.004 2.382.745 4.04 1.093 4.915 1.123.886.03.867-.04 1.155-.243.576-.403 2.22-2.97 5.358-7.158 3.38-4.512 7.335-10.153 10.496-14.725z'/>
-              </svg>
-            </button>
-          </div>
           <div class='flex-1 overflow-y-auto' id='conversation-list'></div>
           <div class='p-2 flex items-center pb-4'>
             <div class='flex-1 message-input-wrapper'>
               <textarea
                 id='message-input'
                 type='text'
-                placeholder='Ask me...'
+                placeholder='Ask Poly about APIs and Events...'
                 onInput='this.parentNode.dataset.messageValue = this.value'
               ></textarea>
             </div>